move scale_lr to base strategy
