+++ .gitignore
@@ -116,8 +116,6 @@ venv.bak/
 *.log.json
 docs/modelzoo_statistics.md
 work_dirs/
-data
-data/
 
 # Pytorch
 *.pth
