+++ mmengine/_strategy/single_device.py
@@ -7,7 +7,7 @@ import torch.nn as nn
 import mmengine
 from mmengine.device import get_device
 from mmengine.model import revert_sync_batchnorm
-from mmengine.optim import BaseOptimWrapper, OptimWrapperDict, _ParamScheduler
+from mmengine.optim import BaseOptimWrapper, _ParamScheduler
 from mmengine.registry import STRATEGIES
 from mmengine.utils import get_git_hash
 from .base import BaseStrategy
@@ -15,23 +15,7 @@ from .base import BaseStrategy
 
 @STRATEGIES.register_module()
 class SingleDeviceStrategy(BaseStrategy):
-    """Strategy for single device training.
-
-    Args:
-        auto_scale_lr (dict, Optional): Config to scale the learning rate
-            automatically. It includes ``base_batch_size`` and ``enable``.
-            ``base_batch_size`` is the batch size that the optimizer lr is
-            based on. ``enable`` is the switch to turn on and off the feature.
-    """
-
-    def __init__(
-        self,
-        *,
-        auto_scale_lr: Optional[dict] = None,
-        **kwargs,
-    ):
-        super().__init__(**kwargs)
-        self.auto_scale_lr = auto_scale_lr
+    """Strategy for single device training."""
 
     def prepare(
         self,
@@ -62,6 +46,8 @@ class SingleDeviceStrategy(BaseStrategy):
                 Defaults to False. Requires PyTorch>=2.0.
             dispatch_kwargs (dict, optional): Kwargs to be passed to other
                 methods of Strategy. Defaults to None.
+                If ``accumulative_counts`` is set in ``optim_wrapper``, you
+                need to provide ``max_iters`` in ``dispatch_kwargs``.
         """
         if dispatch_kwargs is not None:
             self.dispatch_kwargs.update(dispatch_kwargs)
@@ -87,6 +73,18 @@ class SingleDeviceStrategy(BaseStrategy):
         if optim_wrapper is not None:
             self._scale_lr()
 
+            accumulative_counts = getattr(self.optim_wrapper,
+                                          '_accumulative_counts', 1)
+            if accumulative_counts > 1:
+                if 'max_iters' not in self.dispatch_kwargs:
+                    raise ValueError(
+                        '"max_iters" must be specified because '
+                        '"accumulative_counts" was set as '
+                        f'{accumulative_counts} which is greater than 1.')
+
+                self.optim_wrapper.initialize_count_status(  # type: ignore
+                    self.model, 0, self.dispatch_kwargs['max_iters'])
+
         return return_items[0] if len(return_items) == 1 else return_items
 
     def _wrap_model(self, model: nn.Module) -> nn.Module:
@@ -110,58 +108,6 @@ class SingleDeviceStrategy(BaseStrategy):
             'BatchNormXd layers if they are used.')
         model = revert_sync_batchnorm(model)
         return model
-
-    def _scale_lr(self) -> None:
-        """Automatically scaling learning rate in training according to the
-        ratio of ``base_batch_size`` in ``autoscalelr_cfg`` and real batch
-        size.
-
-        It scales the learning rate linearly according to the
-        `paper <https://arxiv.org/abs/1706.02677>`_.
-
-        Note:
-            ``scale_lr`` must be called after building optimizer wrappers
-            and before building parameter schedulers.
-        """
-        if (self.auto_scale_lr is None
-                or not self.auto_scale_lr.get('enable', False)):
-            return None
-
-        assert 'base_batch_size' in self.auto_scale_lr, \
-            'Lack of `base_batch_size` in `auto_scale_lr`.'
-
-        real_bs = self.world_size * self.dispatch_kwargs[
-            'train_micro_batch_size_per_gpu']
-        base_bs = self.auto_scale_lr['base_batch_size']
-        ratio = float(real_bs) / float(base_bs)
-        self.logger.info(f'LR is set based on batch size of {base_bs} '
-                         f'and the current batch size is {real_bs}. '
-                         f'Scaling the original LR by {ratio}.')
-
-        def _is_built(schedulers):
-            if isinstance(schedulers, dict):
-                return False if 'type' in schedulers else any(
-                    _is_built(s) for s in schedulers.values())
-            if isinstance(schedulers, list):
-                return any(_is_built(s) for s in schedulers)
-            return isinstance(schedulers, _ParamScheduler)
-
-        if _is_built(self.param_schedulers):
-            raise RuntimeError('`scale_lr` should be called before building '
-                               'ParamScheduler because ParamScheduler will '
-                               'store initial lr from optimizer wrappers')
-
-        assert isinstance(self.optim_wrapper, BaseOptimWrapper), \
-            '`scale_lr should be called after building OptimWrapper'
-
-        if isinstance(self.optim_wrapper, OptimWrapperDict):
-            wrappers = list(self.optim_wrapper.values())
-        else:
-            wrappers = [self.optim_wrapper]  # type: ignore
-
-        for wrapper in wrappers:
-            for group in wrapper.optimizer.param_groups:
-                group['lr'] = group['lr'] * ratio
 
     def load_checkpoint(
         self,
@@ -209,11 +155,6 @@ class SingleDeviceStrategy(BaseStrategy):
         state_dict = checkpoint.pop('state_dict')
         self.load_model_state_dict(
             state_dict, strict=strict, revise_keys=revise_keys)
-
-        if hasattr(self, 'optim_wrapper'):
-            # Initiate inner count of `optim_wrapper`.
-            self.optim_wrapper.initialize_count_status(  # type: ignore
-                self.model, 0, self.dispatch_kwargs['max_iters'])
 
         return checkpoint
 
@@ -278,9 +219,17 @@ class SingleDeviceStrategy(BaseStrategy):
         cur_iter = checkpoint['meta']['iter']
 
         if hasattr(self, 'optim_wrapper'):
-            # Initiate inner count of `optim_wrapper`.
-            self.optim_wrapper.initialize_count_status(  # type: ignore
-                self.model, cur_iter, self.dispatch_kwargs['max_iters'])
+            accumulative_counts = getattr(self.optim_wrapper,
+                                          '_accumulative_counts', 1)
+            if accumulative_counts > 1:
+                if 'max_iters' not in self.dispatch_kwargs:
+                    raise ValueError(
+                        '"max_iters" must be specified because '
+                        '"accumulative_counts" was set as '
+                        f'{accumulative_counts} which is greater than 1.')
+                    # Initiate inner count of `optim_wrapper`.
+                self.optim_wrapper.initialize_count_status(  # type: ignore
+                    self.model, cur_iter, self.dispatch_kwargs['max_iters'])
 
         return checkpoint
