+++ mmengine/runner/_flexible_runner.py
@@ -23,7 +23,7 @@ from mmengine.optim import OptimWrapper, OptimWrapperDict, _ParamScheduler
 from mmengine.registry import (DATA_SAMPLERS, DATASETS, EVALUATOR, FUNCTIONS,
                                HOOKS, LOG_PROCESSORS, LOOPS, RUNNERS,
                                STRATEGIES, VISUALIZERS, DefaultScope)
-from mmengine.utils import digit_version, is_seq_of
+from mmengine.utils import digit_version
 from mmengine.utils.dl_utils import TORCH_VERSION
 from mmengine.visualization import Visualizer
 from .base_loop import BaseLoop
@@ -65,8 +65,8 @@ class FlexibleRunner:
             a dict used for build a model.
 
     Kwargs:
-        work_dir (str): The working directory to save checkpoints. The logs
-            will be saved in the subdirectory of `work_dir` named
+        work_dir (str, optional): The working directory to save checkpoints.
+            The logs will be saved in the subdirectory of `work_dir` named
             :attr:`timestamp`. Defaults to 'work_dir'.
         experiment_name (str, optional): Name of current experiment. If not
             specified, timestamp will be used as ``experiment_name``.
@@ -313,12 +313,6 @@ class FlexibleRunner:
                 'param_scheduler should be None when optim_wrapper is None, '
                 f'but got {param_scheduler}')
 
-        # Parse `param_scheduler` to a list or a dict. If `optim_wrapper` is a
-        # `dict` with single optimizer, parsed param_scheduler will be a
-        # list of parameter schedulers. If `optim_wrapper` is
-        # a `dict` with multiple optimizers, parsed `param_scheduler` will be
-        # dict with multiple list of parameter schedulers.
-        self._check_scheduler_cfg(param_scheduler)
         self.param_schedulers = param_scheduler
 
         val_related = [val_dataloader, val_cfg, val_evaluator]
@@ -424,7 +418,7 @@ class FlexibleRunner:
         cfg = copy.deepcopy(cfg)
         runner = cls(
             model=cfg['model'],
-            work_dir=cfg['work_dir'],
+            work_dir=cfg.get('work_dir', 'work_dirs'),
             experiment_name=cfg.get('experiment_name'),
             train_dataloader=cfg.get('train_dataloader'),
             optim_wrapper=cfg.get('optim_wrapper'),
@@ -636,13 +630,11 @@ class FlexibleRunner:
                 if strategy is None:
                     strategy = dict(type='DDPStrategy')
 
-                if strategy['type'] == 'DDPStrategy':
-                    strategy.setdefault('auto_scale_lr', self._auto_scale_lr)
-
             assert isinstance(strategy, dict)
 
             strategy.setdefault('work_dir', self._work_dir)
             strategy.setdefault('experiment_name', experiment_name)
+            strategy.setdefault('auto_scale_lr', self._auto_scale_lr)
 
             env_kwargs = dict(
                 launcher=launcher,
@@ -1599,82 +1591,6 @@ class FlexibleRunner:
         else:
             filename = f'{self.timestamp}.py'
         self.cfg.dump(osp.join(self.work_dir, filename))
-
-    def _check_scheduler_cfg(
-        self,
-        param_scheduler: Optional[Union[dict, list, _ParamScheduler]],
-    ) -> None:
-        """Parse `param_scheduler` to a list of parameter schedulers, or a
-        `dict` of which each value is a list of parameter schedulers.
-
-        If only one optimizer is used, the parsed config should be a
-        list of parameter scheduler configs or instances. If multiple
-        optimizers are used, the parsed config should be `dict`.
-        Its key should be consistent with the optimizer `dict` and its value
-        should be a list of parameter scheduler configs or instances. See
-        :meth:`build_param_scheduler` for more details.
-
-        Examples:
-            >>> # valid scheduler:
-            >>> # empty scheduler
-            >>> scheduler = None
-            >>> # Single scheduler
-            >>> scheduler = dict(type='MultiStepLR', milestones=[1, 2])
-            >>> # Single list schedulers
-            >>> scheduler = [dict(type='MultiStepLR', milestones=[1, 2]),
-            >>>              dict(type='MultiStepLR', milestones=[2, 3])]
-            >>> # `dict` of schedulers
-            >>> scheduler = dict(linear1=dict(type='MultiStepLR', milestones=[1, 2]),
-            >>>                  linear2=dict(type='MultiStepLR', milestones=[1, 2]))
-            >>> # `dict` of `list` of schedulers
-            >>> scheduler = dict(linear1=[dict(type='MultiStepLR', milestones=[1, 2])],
-            >>>                  linear2=[dict(type='MultiStepLR', milestones=[1, 2])])
-            >>> # Single built scheduler
-            >>> from mmengine.optim import MultiStepLR
-            >>> scheduler = MultiStepLR(milestones=[1, 2], optimizer=optimizer)
-            >>> # Single built list schedulers
-            >>> scheduler = [MultiStepLR(milestones=[1, 2], optimizer=optimizer)]
-            >>> # dict of built scheduler
-            >>> scheduler = dict(linear1=MultiStepLR(milestones=[1, 2], optimizer=optimizer),
-            >>>                  linear2=MultiStepLR(milestones=[1, 2], optimizer=optimizer))
-            >>> # dict of built list schedulers
-            >>> scheduler = dict(linear1=[MultiStepLR(milestones=[1, 2], optimizer=optimizer)],
-            >>>                  linear2=[MultiStepLR(milestones=[1, 2], optimizer=optimizer)])
-
-        Args:
-            param_scheduler (dict or list): The original parameter scheduler.
-        """  # noqa: E501
-        if param_scheduler is None:
-            return
-        if isinstance(param_scheduler, _ParamScheduler):
-            return
-        if is_seq_of(param_scheduler, _ParamScheduler):
-            return
-
-        if is_seq_of(param_scheduler, dict):
-            for _param_scheduler in param_scheduler:
-                assert 'type' in _param_scheduler, (
-                    'Each parameter scheduler should contain the key type, '
-                    f'but got {_param_scheduler}')
-        elif isinstance(param_scheduler, dict):
-            if 'type' not in param_scheduler:
-                for key, _param_scheduler in param_scheduler.items():
-                    assert isinstance(
-                        _param_scheduler,
-                        (dict, tuple, list, _ParamScheduler)), (
-                            'Each value of `param_scheduler` should be a '
-                            f'dict or a list, but got {_param_scheduler} with '
-                            f'type {type(_ParamScheduler)}')
-
-        else:
-            raise TypeError(
-                '`param_scheduler` should be a `_ParamScheduler`, `dict`, '
-                f'list or a tuple, but got {type(param_scheduler)}. If '
-                '`param_scheduler` is a list of dict, it means a list of '
-                'scheduler configs for single optimizer. If it is a dict and '
-                'contains key `type`, it means a scheduler config for a '
-                'single optimizer. If it does not contain key `type`, it '
-                'means multiple lists of schedulers for multiple optimizers.')
 
     def _log_env(self) -> None:
         """Logging environment information of the current task.
